mv ua script
